requirements.txt
asgiref==3.7.2
Django==4.2.2
django-enum==1.2.1
django-enumeration==1.0.0
django-enumfield==3.1
djangorestframework==3.14.0
Pillow==9.5.0
psycopg2==2.9.6
pytz==2023.3
sqlparse==0.4.4
tzdata==2023.3
